Move Resource specific Ledger Simulator methods into separate trait
