+++ scrypto-test/src/ledger_simulator/mod.rs
@@ -1,6 +1,7 @@
 mod compile;
 mod inject_costing_err;
 mod ledger_simulator;
+mod ledger_simulator_resource_ext;
 
 pub use compile::*;
 pub use inject_costing_err::*;
